pyproject.toml
[project]
name = "codex-playground"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openai",
    "python-dotenv",
    "openpyxl",
    "pydantic",
    "python-docx",
    "markdown-it-py",
    "mdit-py-plugins",
    "pillow",
]

[dependency-groups]
dev = [
    "pytest>=7",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "scripts"]
markers = [
    "integration: tests requiring real API keys or external services",
]
